Convert to page router to use framer motion exit animation
